Удален модуль pdf_converter.py, заменен на document_processor.py для работы с Markdown-файлами. Обновлены команды и логика в telegram_bot.py для поддержки новой структуры. Изменены ссылки на директории и файлы в README.md. Обновлен requirements.txt, удалена зависимость от docling.
