+++ README.md
@@ -1,15 +1,14 @@
-# Телеграм-бот AI-ассистент с поиском по PDF-файлам
+# Телеграм-бот AI-ассистент с поиском по документам
 
-Данное приложение реализует функционал AI-ассистента с гибридным поиском по PDF-файлам на основе Yandex Cloud ML SDK. Ассистент работает в виде телеграм-бота и использует технологии Yandex Cloud для обработки PDF-документов со сложным форматированием и поиска информации в них.
+Данное приложение реализует функционал AI-ассистента с гибридным поиском по документам на основе Yandex Cloud ML SDK. Ассистент работает в виде телеграм-бота и использует технологии Yandex Cloud для работы с Markdown-документами и поиска информации в них.
 
 ## Особенности
 
-- Конвертирование PDF-файлов в Markdown с помощью библиотеки docling
-- Загрузка Markdown-файлов в Yandex Cloud и создание гибридного поискового индекса
+- Загрузка готовых Markdown-файлов в Yandex Cloud и создание гибридного поискового индекса
 - Сохранение ID поискового индекса для повторного использования без переиндексации
 - Управление сессиями пользователей в телеграм-боте (каждый пользователь общается с AI моделью внутри своего треда)
-- Визуализация прогресса обработки документов с помощью прогресс-бара
-- Возможность отмены длительной обработки документов
+- Визуализация прогресса загрузки документов с помощью прогресс-бара
+- Возможность отмены загрузки документов
 - Просмотр истории диалога и сохранение контекста разговора
 - Принудительное пересоздание индекса без перезапуска бота
 - Отказоустойчивость и обработка ошибок
@@ -51,7 +50,7 @@ DEFAULT_TEMPERATURE=0.5
 DEFAULT_SYSTEM_PROMPT="Ваш системный промпт для модели"
 ```
 
-5. Поместите PDF-файлы, которые хотите использовать для поиска, в папку `./data/docs/`.
+5. Поместите Markdown-файлы, которые хотите использовать для поиска, в папку `./data/md/`.
 
 ## Запуск
 
@@ -64,15 +63,15 @@ python main.py
 
 1. Найдите своего бота в Telegram по его имени.
 2. Используйте команду `/start` для начала общения с ботом.
-3. Задавайте вопросы боту, и он будет искать ответы в загруженных PDF-файлах.
+3. Задавайте вопросы боту, и он будет искать ответы в загруженных Markdown-файлах.
 
 ### Доступные команды
 
 - `/start` - Начать общение с ботом
 - `/help` - Показать справку
 - `/reset` - Сбросить историю разговора
-- `/status` - Получить текущий статус обработки документов
-- `/cancel` - Отменить текущую обработку документов
+- `/status` - Получить текущий статус загрузки документов
+- `/cancel` - Отменить текущую загрузку документов
 - `/history` - Показать историю разговора
 - `/reindex` - Принудительно пересоздать поисковый индекс
 
@@ -80,11 +79,19 @@ python main.py
 
 - `main.py` - основной файл для запуска приложения
 - `telegram_bot.py` - модуль для телеграм-бота
-- `pdf_converter.py` - модуль для конвертирования PDF в Markdown и загрузки в Yandex Cloud
+- `document_processor.py` - модуль для загрузки Markdown-файлов в Yandex Cloud
 - `session_manager.py` - модуль для управления сессиями и тредами пользователей
-- `data/docs/` - папка для хранения PDF-файлов
-- `data/md/` - папка для хранения конвертированных Markdown-файлов
+- `data/md/` - папка для хранения Markdown-файлов
 - `data/index_config.json` - файл с сохраненным ID индекса для повторного использования
+
+## Подготовка документов
+
+Перед использованием приложения подготовьте ваши документы в формате Markdown и поместите их в папку `data/md/`. Убедитесь, что:
+
+1. Файлы имеют расширение `.md`
+2. Файлы содержат корректный Markdown-разметку
+3. Файлы не пустые
+4. Кодировка файлов UTF-8
 
 ## Оптимизация работы с индексом
 
@@ -92,7 +99,7 @@ python main.py
 
 1. При первом запуске создается индекс и его ID сохраняется в файле `data/index_config.json`
 2. При последующих запусках проверяется существование индекса в Yandex Cloud
-3. Если индекс существует, он используется без повторной обработки документов
+3. Если индекс существует, он используется без повторной загрузки документов
 4. Если индекс не найден, автоматически создается новый
 5. Команда `/reindex` позволяет принудительно пересоздать индекс, например, при добавлении новых документов
 
@@ -106,7 +113,7 @@ python main.py
 ## Обработка ошибок
 
 Приложение устойчиво к различным ошибкам:
-- SSL-сертификаты обрабатываются автоматически
-- Ошибки при конвертации отдельных документов не прерывают обработку остальных
+- Ошибки при загрузке отдельных документов не прерывают обработку остальных
 - Возможность работы в режиме обычного ассистента без поиска, если создание индекса не удалось
-- Корректная обработка отмены операций и завершения приложения 
+- Корректная обработка отмены операций и завершения приложения
+- Проверка существования директории с Markdown-файлами
