+++ requirements.txt
@@ -1,5 +1,4 @@
 python-telegram-bot>=20.0
 python-dotenv>=1.0.0
 yandex-cloud-ml-sdk>=0.8.0
-docling>=0.4.0
 asyncio>=3.4.3
